src/pyproject.toml
[tool.poetry]
name = "MovieBot"
version = "0.0.1"
description = "A bot that contains information about movies from TheMovieDB"
authors = ["ManuelJNunez <mjnunez@correo.ugr.es>", "IgnasiCR <ignasicr@correo.ugr.es>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
pyrogram = "^1.0.7"
Flask = "^1.1.2"
peewee = "^3.13.3"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
pathlib = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
